Renaming to lower case due to case insensitivities
