+++ src/chakra/theme/corporate/progress.ts
